+++ index.html
@@ -7,6 +7,12 @@
     <link rel="icon" href="/favicon.ico">
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" />
+    <style>
+        html {
+          color-scheme: light dark;
+          min-height: 100svh;
+        }
+    </style>
     <link
       href="https://fonts.googleapis.com/css2?family=Inter:wght@100..900&family=Quicksand:wght@300..700&display=swap"
       rel="stylesheet"
