+++ src/assets/styles/main.css
@@ -1,4 +1,4 @@
-@import './base.css';
+@import 'base.css';
 
 #app {
   margin: 0 auto;
